pyproject.toml
[tool.poetry]
name = "user_service"
version = "0.1.0"
description = ""
authors = ["ramb1zzy"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.12"
fastapi = { extras = ["all"], version = "0.115.6" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
sqlalchemy = { extras = ["postgresql-asyncpg"], version = "^2.0.37" }
pyjwt = "^2.10.1"
aiofiles = "^24.1.0"
gunicorn = "^23.0.0"
cryptography = "^44.0.0"
bcrypt = "<4.0.0"
black = "^24.10.0"
alembic = "^1.14.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
mypy = "^1.14.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.2"
aiosqlite = "^0.20.0"
pytest-mock = "^3.14.0"
types-passlib = "^1.7.7.20241221"


[tool.mypy]
plugins = ["pydantic.mypy"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
